requirements.test.txt
pytest
pytest-cov
pytest-homeassistant-custom-component==0.13.289
crc==7.1.0
pyserial==3.5